remove __mocks__ directory and mock axios instead
